requirements.txt
django
djangorestframework

# Modules for Statistics
scikit-gstat
numpy==1.25
matplotlib 
gstatsim
gstools